cdc-pipeline/requirements.txt
mysql-connector-python>=8.0.0
clickhouse-connect>=0.7.0
kafka-python>=2.0.0
python-dotenv>=1.0.0
